[CHORE] TestPlugin 분리
