+++ build-logic/src/main/java/plugin/KotlinTestPlugin.kt
@@ -1,0 +1,24 @@
+package plugin
+
+import extention.getVersionCatalog
+import extention.testImplementation
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+import org.gradle.api.artifacts.VersionCatalog
+import org.gradle.api.artifacts.dsl.DependencyHandler
+
+internal class KotlinTestPlugin : Plugin<Project> {
+    override fun apply(target: Project) {
+        with(target) {
+            configureKotlinTest()
+        }
+    }
+}
+
+private fun Project.configureKotlinTest() {
+    applyDependency(dependencies, getVersionCatalog())
+}
+
+private fun applyDependency(handler: DependencyHandler, libs: VersionCatalog) = with(handler) {
+    testImplementation(libs.findLibrary("androidx-junit").get())
+}
