--- build-logic/src/main/java/plugin/TestPlugin.kt
@@ -1,2 +1,0 @@
-package plugin
-
